requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
